# pkg/utils: speed up LoadBackupMeta