### What problem does this PR solve? <!--add issue link with summary if exists-->

Speed up LoadBackupMeta.

Cc https://github.com/pingcap/br/issues/705

Manually tested with 39k tables, it improves fast checksum about 53 times.

<details> <summary>Benchmark results</summary>

```
# master branch
[2021/02/19 14:50:58.551 +08:00] [INFO] [collector.go:60] ["Full backup Success summary: total backup ranges: 117738, total success: 117738, total failed: 0, total take(Full backup time): 2m43.027476154s, total take(real time): 23m31.342884193s, total kv: 117738000, total size(MB): 11298.47, avg speed(MB/s): 69.30"] ["backup checksum"=18.410789857s] ["backup fast checksum"=19m20.755603256s] ["backup total regions"=117824] [BackupTS=423025980821995522] [Size=4608796580]

# this branch
[2021/02/19 16:29:03.509 +08:00] [INFO] [collector.go:60] ["Full backup Success summary: total backup ranges: 117738, total success: 117738, total failed: 0, total take(Full backup time): 2m34.65939439s, total take(real time): 4m22.869289236s, total kv: 117738000, total size(MB): 11298.47, avg speed(MB/s): 73.05"] ["backup checksum"=18.999134215s] ["backup fast checksum"=22.223445159s] ["backup total regions"=117824] [Size=4608796580] [BackupTS=423027824506961921]

# master branch
PASS: schema_test.go:218: testSchemaSuite.BenchmarkLoadBackupMeta1024          1        2805279107 ns/op
PASS: schema_test.go:229: testSchemaSuite.BenchmarkLoadBackupMeta10240         1      428872156932 ns/op
PASS: schema_test.go:207: testSchemaSuite.BenchmarkLoadBackupMeta64          200           9807260 ns/op

# this branch
PASS: schema_test.go:218: testSchemaSuite.BenchmarkLoadBackupMeta1024         50          29454271 ns/op
PASS: schema_test.go:229: testSchemaSuite.BenchmarkLoadBackupMeta10240         1        1150945026 ns/op
PASS: schema_test.go:207: testSchemaSuite.BenchmarkLoadBackupMeta64         1000           1662240 ns/op
```

</details>


### What is changed and how it works?


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Related changes

 - Need to cherry-pick to the release branch

### Release Note

 - Improve backup performance when there are many tables.

<!-- fill in the release note, or just write "No release note" -->